requirements.txt
beautifulsoup4==4.11.1
lxml==4.8.0
numpy==1.21.5
pandas==1.1.5
Pillow==9.5.0
Requests==2.30.0
streamlit==1.22.0
streamlit_option_menu==0.3.2
